refactor: extract ResetSessionDialog into a separate file
